fix: rename wrapper to cbind_api.ts to avoid overwriting index.ts
